\name{CG_karyotype}
\alias{CG_karyotype}


\title{
CompleteGenomics karyotype plotting function
}

\description{
Plots each, color coded by chromosomal coordinate, chromosome against a 
background of the complete genome.
}

\usage{
CG_karyotype(chr,start,end,int,ai,name,xlim,ylim)
}

\arguments{
  	\item{chr}{					
  	Chromosomes name of segm object, as generated from somaticCnvSegmentsNonDiplod file.
	}
  	\item{start}{				
	Start position of segment.
	}
	\item{end}{					
  	End position of segment.	
	}
	\item{int}{					
  	Ratio between relative coverage and segment length.
	}
	\item{ai}{	
  	Allelic imbalance.
	}
	\item{name}{
  	Default is ''.
  	Plot generated as "name_karyotype.png" in your working directory.
	}
  	\item{xlim}{
  	Default is c(0,2.4). Limit of x axis.
	}
  	\item{ylim}{
  	Default is c(0,1). Limit of y axis.
	}
}

\details{
Vertical axis: Allelic imbalance. \cr
Horizontal axis: Total intensity. \cr

The plot is a overview, for a closer look see the plots generated by CG_KaCh().
}

%\value{
%%  ~Describe the value returned
%%  If it is a LIST, use
%%  \item{comp1 }{Description of 'comp1'}
%%  \item{comp2 }{Description of 'comp2'}
%% ...
%}

%\references{
%% ~put references to the literature/web site here ~
%}

\author{
Markus Mayrhofer, \email{markus.mayrhofer@medsci.uu.se} \cr
Sebastian DiLorenzo, \email{sebastian.dilorenzo@medsci.uu.se}
}
%\note{
%%  ~~further notes~~
%}

%% ~Make other sections like Warning with \section{Warning }{....} ~

\seealso{
\code{\link{patchwork.CG.plot}} \cr
\code{\link{CG_KaCh}}
}

%\examples{
%##---- Should be DIRECTLY executable !! ----
%##-- ==>  Define data, use random,
%##--	or do  help(data=index)  for the standard data sets.
%Run examples of karyotype stuff. or not, maybe just point back to CNA.
%}

% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
%\keyword{ ~kwd1 }
%\keyword{ ~kwd2 }% __ONLY ONE__ keyword per line