\name{ideogram}
\alias{ideogram}
\docType{data}

\title{
Ideogram data for the chromosomes.
}

\description{
Ideogram data for the chromosomes.
}


\usage{ideogram}

\format{
  A data frame with the following 6 variables.
  \describe{
    \item{chr}{Which chromosome}
    \item{length}{Length of chromosome}
    \item{start}{Start of centromere}
    \item{mid}{Middle of centromere}
    \item{end}{End of centromere}
    \item{c}{Easy integer of chromosome number}
  }
}

\details{
Ideogram data for the chromosomes. Used in many functions.
}

%\source{
%  \url{http://www.}
%}

\seealso{
  The functions \code{\link{patchwork.CG.plot}} and \code{\link{patchwork.CG.copynumbers}} plot functions use this data.
}

%\examples{
%}

%\keyword{datasets}